files: remove backend resources, current is frontend only
Removed all the files of "ApiPwaBackend" since it was decided that the
current branch will be reserved fully for frontend.

From now on, the backend will be controlled by other branch upstreaming
the current 'main-dev-emailService', which the name was not confirmed
at this point, but it will be defined later.

Signed-off-by: Pedro Lucas Soares da Silva <pedrolucassoaresds@gmail.com>