fix #3: direct link to cadastro.html shows error
The current script is trying to redirect the user to the root directory,
which is not used when hosted on github pages without a custom domain.

This commit changes the target patch in app.js to redirect to a
directory before instead of root, so the script is able to find
the main index.html.

Signed-off-by: Pedro Lucas Soares da Silva <pedrolucassoaresds@gmail.com>